porkapp/requirements.txt
blinker     == 1.9.0
click        ==8.3.0
colorama     ==0.4.6
Flask        ==3.1.2
itsdangerous ==2.2.0
Jinja2       ==3.1.6
MarkupSafe   ==3.0.3
pip         == 21.2.3
setuptools  == 57.4.0
Werkzeug    == 3.1.3
sqlalchemy==2.0.36
pyinstaller==6.16.0
pywebview==6.1
weasyprint==62.3          
python-escpos==3.1        